# app/static/swagger.yaml
swagger: '2.0'
info:
  title: Mechanic Shop API
  description: REST API for customers, mechanics, inventory, and service tickets.
  version: 1.0.0
host: "mechanic-shop-api-4ig7.onrender.com"
basePath: /
schemes:
  - "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'Use format: Bearer <token>'
paths:
  /customers:
    post:
      tags:
        - Customers
      summary: Create a customer
      description: Creates a new customer account using name, email, and password.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateCustomerPayload'
      responses:
        201:
          description: Customer created
          schema:
            $ref: '#/definitions/Customer'
          examples:
            application/json:
              id: 1
              name: John Doe
              email: john.doe@gmail.com
              service_ticket_ids: []
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        409:
          description: Customer with that email already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: A customer with that email already exists.
        500:
          description: Failed to create customer
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to create customer.
    get:
      tags:
        - Customers
      summary: List customers
      description: Returns a paginated list of customers. If pagination values are invalid, all customers are returned.
      parameters:
        - in: query
          name: page
          type: integer
          required: false
          default: 1
        - in: query
          name: per_page
          type: integer
          required: false
          default: 10
      responses:
        200:
          description: Customer list
          schema:
            type: array
            items:
              $ref: '#/definitions/Customer'
          examples:
            application/json:
              - id: 1
                name: test user
                email: test@example.com
                service_ticket_ids: [1]
  /customers/login:
    post:
      tags:
        - Customers
      summary: Login customer
      description: Authenticates a customer and returns a bearer token and customer profile data.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginCredentials'
      responses:
        200:
          description: Login success
          schema:
            $ref: '#/definitions/CustomerLoginResponse'
          examples:
            application/json:
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              customer:
                id: 1
                name: test user
                email: test@example.com
                service_ticket_ids: [1]
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        401:
          description: Invalid credentials
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Invalid email or password.
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: Get current customer's tickets
      description: Returns all service tickets associated with the authenticated customer token.
      security:
        - BearerAuth: []
      responses:
        200:
          description: Ticket list for authenticated customer
          schema:
            $ref: '#/definitions/CustomerTicketsResponse'
          examples:
            application/json:
              customer_id: 1
              service_tickets:
                - id: 1
                  vin: 1HGCM82633A123456
                  service_date: '2023-08-15'
                  description: Brake pads replacement
                  status: Pending
                  customer_id: 1
                  mechanic_ids: [1]
                  inventory_ids: [1]
        500:
          description: Failed to retrieve service tickets
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to retrieve service tickets.
...
  /inventory:
    post:
      tags:
        - Inventory
      summary: Create an inventory item
      description: Creates a new inventory part record. Requires a valid mechanic bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateInventoryPayload'
      responses:
        201:
          description: Inventory item created
          schema:
            $ref: '#/definitions/InventoryItem'
          examples:
            application/json:
              id: 3
              name: Brake Pads
              price: 29.99
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        500:
          description: Failed to create inventory item
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to create inventory item.

    get:
      tags:
        - Inventory
      summary: List inventory items
      description: Returns a paginated list of inventory items. If pagination values are invalid, all items are returned.
      parameters:
        - in: query
          name: page
          type: integer
          required: false
          default: 1
        - in: query
          name: per_page
          type: integer
          required: false
          default: 10
      responses:
        200:
          description: Inventory list
          schema:
            type: array
            items:
              $ref: '#/definitions/InventoryItem'
          examples:
            application/json:
              - id: 3
                name: Brake Pads
                price: 29.99
  /inventory/{inventory_id}:
    get:
      tags:
        - Inventory
      summary: Get one inventory item
      description: Returns a single inventory item by ID.
      parameters:
        - in: path
          name: inventory_id
          required: true
          type: integer
      responses:
        200:
          description: Inventory item
          schema:
            $ref: '#/definitions/InventoryItem'
          examples:
            application/json:
              id: 3
              name: Brake Pads
              price: 29.99
        404:
          description: Inventory item not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Inventory item not found.
        500:
          description: Failed to update inventory item
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update inventory item.
    put:
      tags:
        - Inventory
      summary: Update an inventory item
      description: Updates an inventory item by ID. Requires a valid mechanic bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: inventory_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateInventoryPayload'
      responses:
        200:
          description: Inventory item updated
          schema:
            $ref: '#/definitions/InventoryItem'
          examples:
            application/json:
              id: 3
              name: Updated Brake Pads
              price: 34.99
        404:
          description: Inventory item not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Inventory item not found.
        500:
          description: Failed to delete inventory item
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to delete inventory item.

    delete:
      tags:
        - Inventory
      summary: Delete an inventory item
      description: Deletes an inventory item by ID. Requires a valid mechanic bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: inventory_id
          required: true
          type: integer
      responses:
        200:
          description: Inventory item deleted
          schema:
            $ref: '#/definitions/MessageResponse'
          examples:
            application/json:
              message: Inventory item 1 deleted.
        404:
          description: Inventory item not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Inventory item not found.
...
definitions:
  LoginCredentials:
    type: object
    properties:
      email:
        type: string
        example: test@example.com
      password:
        type: string
        example: password
    required:
      - email
      - password
  Customer:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: test user
      email:
        type: string
        example: test@example.com
      service_ticket_ids:
        type: array
        items:
          type: integer
        example: [1, 4]
...
  InventoryItem:
    type: object
    properties:
      id:
        type: integer
        example: 3
      name:
        type: string
        example: Brake Pads
      price:
        type: number
        format: float
        example: 29.99
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
        example: 1
      vin:
        type: string
        example: 1HGCM82633A123456
      service_date:
        type: string
        format: date
        example: '2023-08-15'
      description:
        type: string
        example: Brake pads replacement
      status:
        type: string
        example: Pending
      customer_id:
        type: integer
        example: 1
      mechanic_ids:
        type: array
        items:
          type: integer
        example: [1]
      inventory_ids:
        type: array
        items:
          type: integer
        example: [1, 2]
  CreateCustomerPayload:
    type: object
    properties:
      name:
        type: string
        example: John Doe
      email:
        type: string
        example: john.doe@gmail.com
      password:
        type: string
        example: securepassword
    required:
      - name
      - email
      - password
...
  CreateInventoryPayload:
    type: object
    properties:
      name:
        type: string
        example: Brake Pads
      price:
        type: number
        format: float
        example: 29.99
    required:
      - name
      - price
  UpdateInventoryPayload:
    type: object
    properties:
      name:
        type: string
        example: Updated Brake Pads
      price:
        type: number
        format: float
        example: 34.99
...
  CustomerLoginResponse:
    type: object
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      customer:
        $ref: '#/definitions/Customer'
...
  CustomerTicketsResponse:
    type: object
    properties:
      customer_id:
        type: integer
        example: 1
      service_tickets:
        type: array
        items:
          $ref: '#/definitions/ServiceTicket'
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: Resource deleted successfully.
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: Validation failed for one or more fields.